Scrollbar width detection & homepage feature type/margins
Added scrollCompensate function to main.js to provide a more accurate
width  for the adjustColumns function, so that will actually line up
with the CSS's media queries).

Also finessed the type styles on the generic "big" homepage feature
content type, and fixed the right hand margin on the caption box in
those same homepage features.